Extract map to its own file
